Cast fix for the int/uint.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@386288 13f79535-47bb-0310-9956-ffa450edef68

+++ jk/native/common/jk_ajp_common.c
@@ -1389,7 +1389,7 @@ static int ajp_process_callback(jk_msg_buf_t *msg,
              * and the length of the chunk (2 bytes). The remaining part of
              * the message is the chunk.
              */
-            if (len > msg->len - 3) {
+            if (len > (unsigned int)(msg->len - 3)) {
                 jk_log(l, JK_LOG_ERROR,
                        "Chunk length too large. Length of AJP message is %i,"
                        " chunk length is %i.", msg->len, len);
